testar

+++ SensorManager/SensorManager.cpp
@@ -16,11 +16,11 @@ void SensorManager::initializeMPU(){
     mpu->initialize();
 }
 
-void SensorManager::readMPU(int16_t *input) {
+void SensorManager::readMPU(double *input) {
     
     mpu->getMotion(&input[0], &input[1], &input[2], &input[3], &input[4], &input[5]);
     
-    printf("a/g: %6hd %6hd %6hd   %6hd %6hd %6hd\n",input[0],input[1],input[2],input[3],input[4],input[5]);
+    printf("a/g: %f %f %f   %f %f %f\n",input[0],input[1],input[2],input[3],input[4],input[5]);
 }
