+++ SensorManager/testRaw.cpp
@@ -3,7 +3,7 @@
 #include "SensorManager.h"
 #include "I2Cdev.h"
 
-int16_t input[6];
+double input[6];
 
 int main(int argc, const char * argv[])
 {
